# dashboard/public/openapi.yaml
openapi: 3.0.0
info:
  title: Project Overwatch API
  description: |
    Enterprise API for high-level management of ERLC Private Servers.
    
    ### 🔑 Identity Systems
    This API stores identifiers as plain strings and does not enforce a specific format. However, certain features require specific ID types to function correctly:
    
    1. **Shifts & Staff On-Duty**: 
       - **Required Format**: **Clerk User ID** (e.g., `user_2mGj...`).
       - **Why**: The `/staff/on-duty` endpoint resolves shift user IDs via the Clerk API to fetch staff names and Roblox usernames. If you pass a non-Clerk ID, the lookup will fail silently.
    
    2. **Punishments**: 
       - **Recommended Format**: **Roblox UserID** (numerical, e.g., `12345678`).
       - **Why**: The dashboard tracks punishments by Roblox ID. Using a different format will work, but may not integrate correctly with the dashboard UI.
       - **`moderatorId`**: Accepts any string. Dashboard uses Clerk IDs, but you can pass Roblox IDs if preferred.
    
    3. **Player Lists & Logs**: 
       - **Format**: **Roblox UserID** (numerical).
       - **Why**: These are sourced directly from the game server via PRC.
    
    ### 📡 Server Identification
    All server-specific endpoints require a `server` parameter.
    - **Accepts**: Both the **Internal System Name** (e.g., `LACOMM Server A`) and the **Custom Display Name**.
    - **Matching**: Case-insensitive, supports partial matching.
    
    ### 🔒 Authentication
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /staff/on-duty:
    get:
      summary: Get On-Duty Staff
      tags: [Staff]
      description: |
        Returns staff currently in an active shift. 
        **Important**: This endpoint looks up shift `userId` values via the Clerk API. 
        Only shifts started with valid **Clerk User IDs** will return enriched data (name, Roblox username, avatar).
      parameters:
        - name: server
          in: query
          required: true
          description: Server name or custom display name.
          example: "LACOMM Server A"
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OnDutyStaff' }
...
  /players:
    get:
      summary: Get Server Players
      tags: [Server]
      description: Returns players currently in-game. Player IDs are **Roblox UserIDs** (numerical).
      parameters:
        - name: server
          in: query
          required: true
          description: Server name or custom display name.
          example: "LACOMM Server A"
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlayerList' }
...
  /punishments:
    get:
      summary: Get Punishment History
      tags: [Moderation]
      description: Returns punishments. Filter by `userId` (typically a Roblox UserID).
      parameters:
        - name: server
          in: query
          required: true
          description: Server name or custom display name.
          example: "LACOMM Server A"
          schema: { type: string }
        - name: userId
          in: query
          description: Filter by target's stored ID (dashboard uses Roblox UserIDs).
          example: "45032"
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Punishment' }
    post:
      summary: Log a Punishment
      tags: [Moderation]
      description: |
        Creates a punishment record. Both `userId` and `moderatorId` are stored as-is.
        **Dashboard convention**: `userId` = Roblox UserID, `moderatorId` = Clerk User ID.
        The API does not enforce this; use whatever format fits your integration.
      parameters:
        - name: server
          in: query
          required: true
          description: Server name or custom display name.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, moderatorId, type]
              properties:
                userId: 
                  type: string
                  description: Target identifier (any string, dashboard uses Roblox UserID).
                  example: "45032"
                moderatorId: 
                  type: string
                  description: Issuer identifier (any string, dashboard uses Clerk User ID).
                  example: "user_2mGjxRbc123"
                type: 
                  type: string
                  enum: ["Warn", "Kick", "Ban", "Ban Bolo"]
                reason: 
                  type: string
                  example: "RDM and LTAP"
      responses:
        '200':
          description: Success
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: API Key prefixed with `pow_`. Obtain from Settings dashboard.
  schemas:
...
    OnDutyStaff:
      type: object
      properties:
        userId: { type: string, description: Clerk User ID. }
        name: { type: string, description: Display name from Clerk. }
        robloxUsername: { type: string, description: Linked Roblox username. }
        robloxId: { type: string, nullable: true, description: Roblox UserID (numerical). }
        imageUrl: { type: string, description: Clerk profile picture URL. }
        shiftStart: { type: string, format: date-time }
...
    PlayerList:
      type: object
      properties:
        serverId: { type: string }
        serverName: { type: string }
        playerCount: { type: integer }
        players:
          type: array
          items: { $ref: '#/components/schemas/Player' }
        timestamp: { type: string, format: date-time }
    Player:
      type: object
      properties:
        name: { type: string, description: Roblox Username. }
        id: { type: string, description: Roblox UserID (numerical). }
        team: { type: string }
        permission: { type: string, description: PRC permission level. }
        vehicle: { type: string, nullable: true }
        callsign: { type: string, nullable: true }
...
    Punishment:
      type: object
      properties:
        id: { type: string }
        serverId: { type: string }
        userId: { type: string, description: Target ID (dashboard uses Roblox UserID). }
        moderatorId: { type: string, description: Issuer ID (dashboard uses Clerk User ID). }
        type: { type: string }
        reason: { type: string, nullable: true }
        resolved: { type: boolean }
        createdAt: { type: string, format: date-time }